fix: move system exec to its own module
