+++ Makefile
@@ -6,15 +6,20 @@
 #    By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/01/19 22:38:45 by itaureli          #+#    #+#              #
-#    Updated: 2022/05/07 21:52:40 by vwildner         ###   ########.fr        #
+#    Updated: 2022/05/08 02:40:26 by vwildner         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
-SRC		=	minishell.c readline.c wip_lexer.c print_error.c execute.c utils.c
+SRC		= readline.c wip_lexer.c print_error.c execute.c utils.c execute_system.c
+
+SRC_MAIN = $(SRC)
+SRC_MAIN += minishell.c
 
 SRC_PATH = ./src
 
-SRCS = $(addprefix $(SRC_PATH)/,$(SRC))
+SRCS_NO_ENTRY = $(addprefix $(SRC_PATH)/,$(SRC))
+
+SRCS = $(addprefix $(SRC_PATH)/,$(SRC_MAIN))
 
 OBJECTS		=	${SRCS:.c=.o}
 
@@ -108,7 +113,7 @@ $(BUILTINS_OBJECTS_PATH)/%.o: $(BUILTINS_SOURCES_PATH)/%.c $(BUILTINS_HEADER)
 test:
 	@$(CC) tests/test_unit.c \
 	libs/libft/*.c libs/builtins/*.c \
-	src/wip_lexer.c src/readline.c src/print_error.c src/signal.c src/execute.c \
+	$(SRCS_NO_ENTRY) \
 	-lreadline -lrt -lm \
 	-o test_unit && ./test_unit
