+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: vwildner <vwildner@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/30 17:04:35 by mfrasson          #+#    #+#             */
-/*   Updated: 2022/05/08 01:02:32 by vwildner         ###   ########.fr       */
+/*   Updated: 2022/05/08 02:26:18 by vwildner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,6 +46,7 @@ char	**parse_input(char *input_text);
 int		pwd(void);
 void	signal_handler(int signal_number);
 
+int		system_exec(t_command *cmd);
 int		execute(t_command *cmd);
 
 void	atexit_clean(void *data);
